requirements.txt
gradio
torch
diffusers
safetensors
pillow
openai
cohere
anthropic
click
llama_cpp_python